moved Header.js inside components folder
